qr_generator/requirements.txt
PyQt6
Pillow
qrcode[pil]
